统计修改

+++ dataanalysis/src/main/java/com/aizhixin/cloud/dataanalysis/analysis/domain/AvgAgeDomain.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class AvgAgeDomain {
     Integer age;
-    Double score;
+    Integer score;
 }
