+++ dataanalysis/src/main/java/com/aizhixin/cloud/dataanalysis/analysis/domain/AvgDomain.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class AvgDomain {
     String name;
-    Double score;
+    Integer score;
 }
