+++ dataanalysis/src/main/java/com/aizhixin/cloud/dataanalysis/analysis/domain/AvgNjDomain.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class AvgNjDomain {
     String nj;
-    Double score;
+    Integer score;
 }
