+++ dataanalysis/src/main/java/com/aizhixin/cloud/dataanalysis/analysis/domain/AvgSexDomain.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class AvgSexDomain {
     String sex;
-    Double score;
+    Integer score;
 }
